Completed Module-2 and 3 of Pluralsight's EntityFramework Core 2: mappings. I could have used the solution from the prior course, but I decided nit would be better to take a raw solution and build it up almost from scratch. I am using a new seeder class to populate the database with information.
On a twist to the Samurai theme Julie Lerman uses for this course, I also thought it better to go with a theme that I know about WWII history.   So I am using WWII pacific battles and well known American participants.   It helps me verify the database is updating properly.